Move generationStatus class into generation folder
