+++ src/extension/src/utils/generation/generators/AppServiceGenerator.ts
@@ -2,7 +2,7 @@ import { AzureServices } from "../../../azure/azureServices";
 import { MESSAGES } from "../../../constants/messages";
 import { TelemetryEventName } from "../../../constants/telemetry";
 import { IAppService, IGenerationData, IService, SERVICE_TYPE } from "../../../types/generationTypes";
-import { sendGenerationStatus, GenerationItemStatus } from "../../generationStatus";
+import { sendGenerationStatus, GenerationItemStatus } from "../generationStatus";
 import { Logger } from "../../logger";
 import { DeployedServiceStatus } from "../GenerationService";
 import { IGenerator } from "../IGenerator";
